+++ src/completion-provider/template-context-provider.ts
@@ -1,43 +1,39 @@
 import { join } from 'path';
 import uniqueBy from '../utils/unique-by';
 import { readFileSync, existsSync } from 'fs';
+import { CompletionItem, CompletionItemKind } from 'vscode-languageserver';
+
 import {
-    CompletionItem,
-    CompletionItemKind
-} from 'vscode-languageserver';
+  isModuleUnificationApp,
+  safeWalkSync,
+  podModulePrefixForRoot
+} from '../utils/layout-helpers';
 
-import { 
-  isModuleUnificationApp, podModulePrefixForRoot, safeWalkSync
-} from './template-completion-provider';
-// const debug = false;
-// const fs = require('fs');
-// const util = require('util');
-// const log_file = fs.createWriteStream(__dirname + '/debug.log', {flags : 'w'});
+import { log } from '../utils/logger';
 
-// console.log = debug ? function(...args: any[]) {
-//   const output = args.map((a: any) => {
-//     return JSON.stringify(a);
-//   }).join(' ');
-//   log_file.write('----------------------------------------' + '\r\n');
-//   log_file.write(util.format(output) + '\r\n');
-//   log_file.write('----------------------------------------' + '\r\n');
-// } : function() {};
+import {
+  extractComponentInformationFromMeta,
+  processJSFile,
+  processTemplate
+} from 'ember-meta-explorer';
 
-import { extractComponentInformationFromMeta, processJSFile, processTemplate }  from 'ember-meta-explorer';
-
-export function templateContextLookup(root: string, currentFilePath: string, templateContent: string) {
-    console.log('templateContextLookup', root, currentFilePath, templateContent);
-    const nameParts = currentFilePath.split('/components/');
-    if (nameParts.length !== 2) {
-        return [];
-    }
-    let componentName = nameParts[1].split('.')[0];
-    if (componentName.endsWith('/component')) {
-      componentName = componentName.replace('/component', '');
-    } else if (componentName.endsWith('/template')) {
-      componentName = componentName.replace('/template', '');
-    }
-    return componentsContextData(root, componentName, templateContent);
+export function templateContextLookup(
+  root: string,
+  currentFilePath: string,
+  templateContent: string
+) {
+  log('templateContextLookup', root, currentFilePath, templateContent);
+  const nameParts = currentFilePath.split('/components/');
+  if (nameParts.length !== 2) {
+    return [];
+  }
+  let componentName = nameParts[1].split('.')[0];
+  if (componentName.endsWith('/component')) {
+    componentName = componentName.replace('/component', '');
+  } else if (componentName.endsWith('/template')) {
+    componentName = componentName.replace('/template', '');
+  }
+  return componentsContextData(root, componentName, templateContent);
 }
 
 function getComponentsScriptsFolder(root: string) {
@@ -74,8 +70,12 @@ function getPoddedComponentsFileLocationPath(root: string, filePath: string) {
   }
 }
 
-function componentsContextData(root: string, postfix: string, templateContent: string): CompletionItem[] {
-    console.log('templateContextLookup', root, postfix, templateContent);
+function componentsContextData(
+  root: string,
+  postfix: string,
+  templateContent: string
+): CompletionItem[] {
+  log('templateContextLookup', root, postfix, templateContent);
   const jsPaths = safeWalkSync(getComponentsScriptsFolder(root), {
     directories: false,
     globs: [
@@ -88,67 +88,74 @@ function componentsContextData(root: string, postfix: string, templateContent: s
 
   const jsPodsPaths = safeWalkSync(getPoddedComponentsScriptsFolder(root), {
     directories: false,
-    globs: [
-      `**/**/${postfix}/component.js`,
-      `**/**/${postfix}/component.ts`
-    ]
+    globs: [`**/**/${postfix}/component.js`, `**/**/${postfix}/component.ts`]
   });
 
-  console.log('jsPaths', jsPaths);
-  const infoItems = [].concat.apply([], [...jsPodsPaths, ...jsPaths].filter((fileName: string) => {
-    return !!fileName;
-  }).map((filePath: string) => {
-    const fileLocation = getComponentFileLocationPath(root, filePath);
-    const podFileLocation = getPoddedComponentsFileLocationPath(root, filePath);
-    console.log('fileLocation', fileLocation);
-    let fileContent = '';
-    if (existsSync(fileLocation)) {
-      fileContent = readFileSync(fileLocation, { encoding: 'utf8' });
-    } else if (podFileLocation && existsSync(podFileLocation)) {
-      fileContent = readFileSync(podFileLocation, { encoding: 'utf8'});
-    } else {
-      return null;
-    }
-    console.log('fileContent', fileContent);
-    try {
-        const jsMeta = processJSFile(fileContent, filePath);
-        console.log('jsMeta', jsMeta);
-        return jsMeta;
-    } catch (e) {
-        console.log('error', e);
-        return null;
-    }
-  }));
+  log('jsPaths', jsPaths);
+  const infoItems = [].concat.apply(
+    [],
+    [...jsPodsPaths, ...jsPaths]
+      .filter((fileName: string) => {
+        return !!fileName;
+      })
+      .map((filePath: string) => {
+        const fileLocation = getComponentFileLocationPath(root, filePath);
+        const podFileLocation = getPoddedComponentsFileLocationPath(
+          root,
+          filePath
+        );
+        log('fileLocation', fileLocation);
+        let fileContent = '';
+        if (existsSync(fileLocation)) {
+          fileContent = readFileSync(fileLocation, { encoding: 'utf8' });
+        } else if (podFileLocation && existsSync(podFileLocation)) {
+          fileContent = readFileSync(podFileLocation, { encoding: 'utf8' });
+        } else {
+          return null;
+        }
+        log('fileContent', fileContent);
+        try {
+          const jsMeta = processJSFile(fileContent, filePath);
+          log('jsMeta', jsMeta);
+          return jsMeta;
+        } catch (e) {
+          log('error', e);
+          return null;
+        }
+      })
+  );
 
   let templateInfo: any = null;
   try {
     templateInfo = processTemplate(templateContent);
   } catch (e) {
-    console.log('templateError', e);
+    log('templateError', e);
   }
   infoItems.push(templateInfo);
-  console.log('infoItems', infoItems);
+  log('infoItems', infoItems);
 
-  const meta: any = infoItems.filter((item: any) => item !== null).reduce((result: any, it: any) => {
-    console.log('it', it);
-    Object.keys(it).forEach(name => {
-      if (name in result) {
-        result[name] = result[name].concat(it[name]);
-      } else {
-        result[name] = it[name].slice(0);
-      }
-    });
-    return result;
-  }, {});
+  const meta: any = infoItems
+    .filter((item: any) => item !== null)
+    .reduce((result: any, it: any) => {
+      log('it', it);
+      Object.keys(it).forEach(name => {
+        if (name in result) {
+          result[name] = result[name].concat(it[name]);
+        } else {
+          result[name] = it[name].slice(0);
+        }
+      });
+      return result;
+    }, {});
   const items: any = [];
-  console.log('meta', meta);
+  log('meta', meta);
   let contextInfo: any = {};
   try {
-   contextInfo = extractComponentInformationFromMeta(meta);
+    contextInfo = extractComponentInformationFromMeta(meta);
   } catch (e) {
-    console.log('contextInforError', e);
+    log('contextInforError', e);
   }
-  console.log('contextInfo', contextInfo);
+  log('contextInfo', contextInfo);
   function localizeName(name: string) {
     if (name.startsWith('this.')) {
       return name;
@@ -164,7 +171,7 @@ function componentsContextData(root: string, postfix: string, templateContent: s
     items.push({
       kind: CompletionItemKind.Property,
       label: localizeName(name),
-      detail: propName,
+      detail: propName
     });
   });
   contextInfo.jsComputeds.forEach((propName: string) => {
@@ -172,7 +179,7 @@ function componentsContextData(root: string, postfix: string, templateContent: s
     items.push({
       kind: CompletionItemKind.Property,
       label: localizeName(name),
-      detail: 'ComputedProperty: ' + propName,
+      detail: 'ComputedProperty: ' + propName
     });
   });
   contextInfo.jsFunc.forEach((propName: string) => {
@@ -180,7 +187,7 @@ function componentsContextData(root: string, postfix: string, templateContent: s
     items.push({
       kind: CompletionItemKind.Function,
       label: localizeName(name),
-      detail: 'Function: ' + propName,
+      detail: 'Function: ' + propName
     });
   });
   contextInfo.hbsProps.forEach((propName: string) => {
@@ -188,7 +195,7 @@ function componentsContextData(root: string, postfix: string, templateContent: s
     items.push({
       kind: CompletionItemKind.Function,
       label: name,
-      detail: 'Template Property: ' + propName,
+      detail: 'Template Property: ' + propName
     });
   });
   // contextInfo.api.actions.forEach((propName: string) => {
