+++ src/definition-provider.ts
@@ -10,228 +10,32 @@ import {
 } from 'vscode-languageserver';
 
 import { parse } from 'babylon';
-
+import URI from 'vscode-uri';
 import { toPosition } from './estree-utils';
 import Server from './server';
 import ASTPath from './glimmer-utils';
 import { getExtension } from './utils/file-extension';
+import { isTransformReference, isModelReference } from './utils/ast-helpers';
 import {
-  getProjectAddonsRoots,
-  getPodModulePrefix,
-  isModuleUnificationApp
-} from './completion-provider/template-completion-provider';
-import URI from 'vscode-uri';
+  isTemplatePath,
+  getComponentNameFromURI
+} from './utils/layout-helpers';
+
+import {
+  getAbstractHelpersParts,
+  getAddonPathsForComponentTemplates,
+  getPathsForComponentTemplates,
+  getPathsForComponentScripts
+} from './utils/definition-helpers';
+
 const _ = require('lodash');
 const memoize = require('memoizee');
+const { preprocess } = require('@glimmer/syntax');
 
-const mProjectAddonsRoots = memoize(getProjectAddonsRoots, {
-  length: 1,
-  maxAge: 600000
-});
 const mAddonPathsForComponentTemplates = memoize(
   getAddonPathsForComponentTemplates,
   { length: 2, maxAge: 600000 }
 );
-
-const { preprocess } = require('@glimmer/syntax');
-
-function getAddonPathsForComponentTemplates(
-  root: string,
-  maybeComponentName: string
-) {
-  const roots = mProjectAddonsRoots(root);
-  let existingPaths: string[] = [];
-  let hasValidPath = false;
-  roots.forEach((rootPath: string) => {
-    if (hasValidPath) {
-      return;
-    }
-    const addonPaths: string[][] = [];
-
-    getAbstractComponentScriptsParts(
-      rootPath,
-      'addon',
-      maybeComponentName
-    ).forEach((parts: any) => {
-      addonPaths.push(parts);
-    });
-    getAbstractComponentScriptsParts(
-      rootPath,
-      'app',
-      maybeComponentName
-    ).forEach((parts: any) => {
-      addonPaths.push(parts);
-    });
-    getAbstractComponentTemplatesParts(
-      rootPath,
-      'app',
-      maybeComponentName
-    ).forEach((parts: any) => {
-      addonPaths.push(parts);
-    });
-    getAbstractComponentTemplatesParts(
-      rootPath,
-      'addon',
-      maybeComponentName
-    ).forEach((parts: any) => {
-      addonPaths.push(parts);
-    });
-    getAbstractHelpersParts(rootPath, 'app', maybeComponentName).forEach(
-      (parts: any) => {
-        addonPaths.push(parts);
-      }
-    );
-    getAbstractHelpersParts(rootPath, 'addon', maybeComponentName).forEach(
-      (parts: any) => {
-        addonPaths.push(parts);
-      }
-    );
-
-    const validPaths = addonPaths
-      .map((pathArr: string[]) => {
-        return path.join.apply(path, pathArr.filter((part: any) => !!part));
-      })
-      .filter(fs.existsSync);
-    if (validPaths.length) {
-      hasValidPath = true;
-      existingPaths = validPaths;
-    }
-  });
-
-  const addonFolderFiles = existingPaths.filter(hasAddonFolderInPath);
-  if (addonFolderFiles.length) {
-    return addonFolderFiles;
-  }
-  return existingPaths;
-}
-
-function getPathsForComponentTemplates(
-  root: string,
-  maybeComponentName: string
-): string[] {
-  const podModulePrefix = getPodModulePrefix(root);
-  let podComponentsScriptsParts: string[][] = [];
-  let muComponentsScriptsParts: string[][] = [];
-  let classicComponentsScriptsParts: string[][] = [];
-  if (podModulePrefix) {
-    podComponentsScriptsParts = getAbstractComponentTemplatesParts(
-      root,
-      'app' + path.sep + podModulePrefix,
-      maybeComponentName
-    );
-  }
-  if (isModuleUnificationApp(root)) {
-    muComponentsScriptsParts = getAbstractComponentTemplatesParts(
-      root,
-      'src/ui',
-      maybeComponentName
-    );
-  } else {
-    classicComponentsScriptsParts = getAbstractComponentTemplatesParts(
-      root,
-      'app',
-      maybeComponentName
-    );
-  }
-  const paths = [
-    ...podComponentsScriptsParts,
-    ...muComponentsScriptsParts,
-    ...classicComponentsScriptsParts
-  ].map((pathParts: any) => {
-    return path.join.apply(path, pathParts.filter((part: any) => !!part));
-  });
-  return paths;
-}
-
-function getAbstractComponentScriptsParts(
-  root: string,
-  prefix: string,
-  maybeComponentName: string
-) {
-  return [
-    [root, prefix, 'components', maybeComponentName + '.js'],
-    [root, prefix, 'components', maybeComponentName + '.ts'],
-    [root, prefix, 'components', maybeComponentName, 'component.js'],
-    [root, prefix, 'components', maybeComponentName, 'component.ts']
-  ];
-}
-
-function getAbstractComponentTemplatesParts(
-  root: string,
-  prefix: string,
-  maybeComponentName: string
-) {
-  return [
-    [root, prefix, 'components', maybeComponentName, 'template.hbs'],
-    [root, prefix, 'templates', 'components', maybeComponentName + '.hbs']
-  ];
-}
-
-function getAbstractHelpersParts(
-  root: string,
-  prefix: string,
-  maybeComponentName: string
-) {
-  return [
-    [root, prefix, 'helpers', `${maybeComponentName}.js`],
-    [root, prefix, 'helpers', `${maybeComponentName}.ts`]
-  ];
-}
-
-function getPathsForComponentScripts(
-  root: string,
-  maybeComponentName: string
-): string[] {
-  const podModulePrefix = getPodModulePrefix(root);
-  let podComponentsScriptsParts: string[][] = [];
-  let muComponentsScriptsParts: string[][] = [];
-  let classicComponentsScriptsParts: string[][] = [];
-  if (podModulePrefix) {
-    podComponentsScriptsParts = getAbstractComponentScriptsParts(
-      root,
-      'app/' + podModulePrefix,
-      maybeComponentName
-    );
-  }
-  if (isModuleUnificationApp(root)) {
-    muComponentsScriptsParts = getAbstractComponentScriptsParts(
-      root,
-      'src/ui',
-      maybeComponentName
-    );
-  } else {
-    classicComponentsScriptsParts = getAbstractComponentScriptsParts(
-      root,
-      'app',
-      maybeComponentName
-    );
-  }
-  const paths = [
-    ...muComponentsScriptsParts,
-    ...podComponentsScriptsParts,
-    ...classicComponentsScriptsParts
-  ].map((pathParts: any) => {
-    return path.join.apply(path, pathParts.filter((part: any) => !!part));
-  });
-  return paths;
-}
-
-function getComponentNameFromURI(root: string, uri: string) {
-  let fileName = uri.replace('file://', '').replace(root, '');
-  let splitter = fileName.includes(path.sep + '-components' + path.sep)
-    ? '/-components/'
-    : '/components/';
-  let maybeComponentName = fileName
-    .split(path.sep)
-    .join('/')
-    .split(splitter)[1];
-  if (maybeComponentName.endsWith('/template.hbs')) {
-    maybeComponentName = maybeComponentName.replace('/template.hbs', '');
-  } else if (isTemplatePath(maybeComponentName)) {
-    maybeComponentName = maybeComponentName.replace('.hbs', '');
-  }
-  return maybeComponentName;
-}
 
 export default class DefinitionProvider {
   constructor(private server: Server) {}
@@ -570,46 +374,6 @@ export default class DefinitionProvider {
   }
 }
 
-function isModelReference(astPath: ASTPath): boolean {
-  let node = astPath.node;
-  if (node.type !== 'StringLiteral') {
-    return false;
-  }
-  let parent = astPath.parent;
-  if (
-    !parent ||
-    parent.type !== 'CallExpression' ||
-    parent.arguments[0] !== node
-  ) {
-    return false;
-  }
-  let identifier =
-    parent.callee.type === 'Identifier'
-      ? parent.callee
-      : parent.callee.property;
-  return identifier.name === 'belongsTo' || identifier.name === 'hasMany';
-}
-
-function isTransformReference(astPath: ASTPath): boolean {
-  let node = astPath.node;
-  if (node.type !== 'StringLiteral') {
-    return false;
-  }
-  let parent = astPath.parent;
-  if (
-    !parent ||
-    parent.type !== 'CallExpression' ||
-    parent.arguments[0] !== node
-  ) {
-    return false;
-  }
-  let identifier =
-    parent.callee.type === 'Identifier'
-      ? parent.callee
-      : parent.callee.property;
-  return identifier.name === 'attr';
-}
-
 function pathsToLocations(...paths: string[]): Location[] {
   return paths.filter(fs.existsSync).map(modulePath => {
     return Location.create(
@@ -617,14 +381,6 @@ function pathsToLocations(...paths: string[]): Location[] {
       Range.create(0, 0, 0, 0)
     );
   });
-}
-
-function isTemplatePath(filePath: string) {
-  return filePath.endsWith('.hbs');
-}
-
-function hasAddonFolderInPath(name: string) {
-  return name.includes(path.sep + 'addon' + path.sep);
 }
 
 // function getLineFromText(text: string, line: number) {
